requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3
python-dotenv==1.0.1
weaviate-client==4.9.4
requests==2.32.5
gradio==5.7.1
